.......... [ZBX-6126] updated to the latest /trunk
git-svn-id: svn://svn.zabbix.com/branches/dev@33133 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-6126/frontends/php/api/classes/CAction.php
@@ -1056,10 +1056,10 @@ class CAction extends CZBXAPI {
 
 						$diff = zbx_array_diff($operation['opcommand_hst'], $operationDb['opcommand_hst'], 'hostid');
 						$opcommandHstCreate = array_merge($opcommandHstCreate, $diff['first']);
-						foreach ($diff['second'] as $omhst) {
+						$opCommandHostIds = zbx_objectValues($diff['second'], 'opcommand_hstid');
+						if ($opCommandHostIds) {
 							DB::delete('opcommand_hst', array(
-								'hostid' => $omhst['hostid'],
-								'operationid' => $operation['operationid']
+								'opcommand_hstid' => $opCommandHostIds
 							));
 						}
 					}
